requirements.txt
alembic==1.8.1
Babel==2.10.3
Fabric==2.7.1
Flask==2.1.3
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
moment==0.12.1
python_dateutil==2.8.2
SQLAlchemy==1.4.39
WTForms==3.0.1
